refactor: CSV Strategies in different files and helper should use this strategies.
